move functions from Button.jsx to index.js
