+++ src/components/UI/Button/Button.jsx
@@ -2,11 +2,8 @@ import React, { useState } from "react";
 import "./button.css";
 import { Link } from "react-router-dom";
 import { useTranslation } from "react-i18next";
-import { ref, set, push } from "firebase/database";
-import { onValue } from "firebase/database";
-import { db } from "../../../index";
 import arrowleft from "../../../images/icons/arrowleft.png";
-import { getTotalQuestionsNumb } from "../../../index";
+import { sendDataDb, highlightPreviousAnswer } from "../../../index";
 import { StyledButton, StyledArticle, StyledButtonBack, StyledImg, StyledSpan, StyledButtonQuiz } from "./Button.Styled";
 
 
@@ -16,11 +13,9 @@ const Button = ({
   setCurrentQuestionNumb,
 }) => {
   const { t } = useTranslation();
-  const answersItem = document.querySelectorAll('#answersAll ul li');
 
   const [uniqueIdUser, setUniqueIdUser] = useState(`${Date.now()}`);
   const [questionStartTime, setQuestionStartTime] = useState(Date.now());
-  let timeQuestion;
 
   const btnHomePage = () => {
     return (
@@ -38,62 +33,11 @@ const Button = ({
   };
 
   const btnsQuizPage = () => {
-    let referenceUserAnswers = ref(
-      db,
-      `user${uniqueIdUser}/answer${currentQuestionNumb}`
-    );
-
+    const answersItem = document.querySelectorAll('#answersAll ul li');
+    let timeQuestion;
     const getTimeQuestion = () => {
       timeQuestion = ((Date.now() - questionStartTime) / 1000).toFixed(2);
       setQuestionStartTime(Date.now());
-    };
-
-    const setUniqueIdAnswer = () => {
-      const answerId = push(referenceUserAnswers).key;
-      return answerId;
-    };
-
-    const sendDataDb = (
-      currentQuestionNumb,
-      question,
-      userAnswer,
-      theme,
-      timeQuestion
-    ) => {
-      let rightAnswerDb;
-      onValue(
-        ref(db, `questions/question${currentQuestionNumb}/rightAnswer`),
-        (snapshot) => {
-          rightAnswerDb = snapshot.val();
-        }
-      );
-      set(referenceUserAnswers, {
-        id: setUniqueIdAnswer(),
-        question: question,
-        userAnswer: userAnswer,
-        theme: theme,
-        point: rightAnswerDb === userAnswer ? 1 : 0,
-        time: timeQuestion,
-      });
-    };
-
-    const highlightPreviousAnswer = () => {
-        onValue(
-          ref(db, `user${uniqueIdUser}/answer${currentQuestionNumb}`),
-          (snapshot) => {
-             setTimeout(() => {
-              answersItem.forEach((answerItem) => {
-                  if (answerItem.textContent === snapshot.val().userAnswer) {
-                    answerItem.style.border = '2px solid rgb(103, 104, 215)';
-                    answerItem.setAttribute('data-useranswer', true);
-                  }
-                });
-            }, 1);     
-    
-          }, {
-            onlyOnce: true
-          }
-        ); 
     };
 
     return (
@@ -103,7 +47,7 @@ const Button = ({
             onClick={(e) => {
               setCurrentQuestionNumb(--currentQuestionNumb);
               e.target.closest("#BtnBack").style.display = "none";
-              highlightPreviousAnswer();
+              highlightPreviousAnswer(uniqueIdUser, currentQuestionNumb, '#answersAll ul li');
             }}
           >
             <StyledImg src={arrowleft} alt="Кнопка назад" />
@@ -113,33 +57,20 @@ const Button = ({
 
         <StyledButtonQuiz
           onClick={() => {
-            const theme =
-              document.querySelector("#themeQuestion").textContent;
-            const question =
-              document.querySelector("#questionName").textContent;
             getTimeQuestion();
-
             answersItem.forEach((asnwerItem) => {
                 if (asnwerItem.dataset.useranswer) {
-                  getTotalQuestionsNumb().then((totalQuestionsNumb) => {
-                    setCurrentQuestionNumb(
-                      currentQuestionNumb === totalQuestionsNumb
-                        ? totalQuestionsNumb
-                        : currentQuestionNumb + 1
-                    );
-                  });
+                  setCurrentQuestionNumb(currentQuestionNumb + 1);
                   sendDataDb(
                     currentQuestionNumb,
-                    question,
+                    "#questionTitle",
                     asnwerItem.textContent,
-                    theme,
-                    timeQuestion
+                    "#themeQuestion",
+                    timeQuestion,
+                    uniqueIdUser
                   );
                 } else {
                   return false;
-                }
-                if (currentQuestionNumb >= 2) {
-                  document.querySelector("#BtnBack").style.display = "flex";
                 }
               });
           }}
